+++ lib/file_hutch/rails.rb
@@ -2,6 +2,7 @@
 
 require "rails"
 require_relative "rails/attachable"
+require_relative "backfill"
 
 module FileHutch
   def self.deprecator = @deprecator ||= ActiveSupport::Deprecation.new("0.3", "file_hutch")